requirements.txt
PyQt5>=5.15
numpy
pandas
h5py
PyYAML
matplotlib
scikit-learn
torch
torchvision
timm
tqdm
